Add files via upload
This project is a team project, where our goal was to make a java application for a newspaper agency using Agile Methodology. Our team regularly met for meetings the project is setup on Bit-bucket and we are working with the concepts like user stories, branching, merging, code coverage, Junit Testing and notification and much more. This experience has helped me learn, how actual work is done in team using Test-Driven Development and gave me the sense about, how an Agile team works.