HSQL database application both Client and server side
This project has 2 sides to it one is the server side which creates a server using which the project is able to perform certain functions like get, put, post and delete. And the other side which runs on server side and uses it HSQL database in order to manipulate the database. It uses an XMLPullParser which translates the data using getter and setter methods. It also contains a video explanation by me for better understanding.